decouple stackChunkOop and its klass
